# doc/api.yaml
...
openapi: 3.0.0
info:
  title: diha-agency-api
  description: An early version of an API for a travel Agency named "Diha Agency".
  version: 1.0.0
servers:
- url: https://virtserver.swaggerhub.com/sanmandresy/diha_agency_api/1.0.0
  description: SwaggerHub API Auto Mocking
paths:
...
  /hotels/{hotel_id}/rooms:
    get:
      tags:
      - Hotel
      description: get all rooms available
      operationId: getRooms
      parameters:
      - name: hotel_id
        in: path
        description: hotel's id
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: returns a list of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Room'
  /hotels/{hotel_id}/rooms/{room_id}:
    get:
      tags:
      - Hotel
      description: get a specified room with its details
      operationId: getRoomById
      parameters:
      - name: hotel_id
        in: path
        description: hotel's id
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      - name: room_id
        in: path
        description: room's id
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: returns a specific room when its id is provided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedException'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFoundException'
        "429":
          description: Too many requests to the API
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerException'
    put:
      tags:
      - Hotel
      description: modify a specified room's details
      operationId: updateRoomById
      parameters:
      - name: hotel_id
        in: path
        description: hotel's id
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      - name: room_id
        in: path
        description: room's id
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      requestBody:
        description: room to update
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreRoom'
        required: true
      responses:
        "200":
          description: return the updated room
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedException'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFoundException'
        "429":
          description: Too many requests to the API
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerException'
    delete:
      tags:
      - Hotel
      description: delete a specified room
      operationId: deleteRoomById
      parameters:
      - name: hotel_id
        in: path
        description: hotel's id
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      - name: room_id
        in: path
        description: room's id
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: room removed successfully !
          content:
            text/plain:
              schema:
                type: string
                example: Room deleted !
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedException'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceNotFoundException'
        "429":
          description: Too many requests to the API
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerException'
...
components:
  schemas:
...
    Client:
      type: object
      properties:
        id:
          type: integer
        first_name:
          type: string
          example: John
        last_name:
          type: string
          example: Doe
        username:
          type: string
          example: jojoe
        sex:
          type: string
          enum:
          - M
          - F
        birth_date:
          type: string
          format: date
        nationality:
          type: string
          example: Malagasy
        address:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
        registration_date:
          type: string
          format: date
        credit_card_number:
          type: string
...
    Hotel:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          example: La Villa Romantica
        location:
          type: string
          example: Roma
        description:
          type: string
        owner:
          type: string
        number_of_rooms:
          type: integer
        ratings:
          type: integer
          example: 5
        phone:
          type: string
        email:
          type: string
          format: email
...
    PreRoom:
      type: object
      properties:
        ref:
          type: string
          example: Venus
        room_type:
          type: string
          example: double
        status:
          type: string
          example: occupied
        price:
          type: integer
        startingDateBooking:
          type: string
          format: date
        endingDateBooking:
          type: string
          format: date
        client_details:
          allOf:
          - $ref: '#/components/schemas/Client'
    Room:
      type: object
      properties:
        room_id:
          type: integer
          example : 56
        hotel:
          allOf:
            - $ref : "#/components/schemas/Hotel"
        client:
          allOf:
            - $ref: "#/components/schemas/Client"
        price :
          type: integer
        ref:
          type: string
          example: Venus
        room_type:
          type: string
          example: double
        status:
          type: string
          example: occupied
        startingDateBooking:
          type: string
          format: date
        endingDateBooking:
          type: string
          format: date
...
    Exception:
      type: object
      properties:
        type:
          type: string
        message:
          type: string
    BadRequestException:
      example:
        type: BadRequestException
        message: Bad request
      allOf:
      - $ref: '#/components/schemas/Exception'
    NotAuthorizedException:
      example:
        type: NotAuthorizedException
        message: Not authorized
      allOf:
      - $ref: '#/components/schemas/Exception'
    ResourceNotFoundException:
      example:
        type: ResourceNotFoundException
        message: Resource of type <T> identified by <I> not found
      allOf:
      - $ref: '#/components/schemas/Exception'
    TooManyRequestsException:
      example:
        type: TooManyRequestsException
        message: Too many requests
      allOf:
      - $ref: '#/components/schemas/Exception'
    InternalServerException:
      example:
        type: InternalServerException
        message: Unexpected error
      allOf:
      - $ref: '#/components/schemas/Exception'
...